Move Hilt module into api.anilist
